app/requirements.txt
# FastAPI 및 웹 서비스 관련
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# 비정형 데이터 처리
unstructured[pdf]
pypdf2==3.0.1
PyMuPDF==1.23.26  # PDF → 이미지 변환용
Pillow
google-cloud-vision==3.7.2

# 설정 및 데이터 검증
pydantic
pydantic-settings
typing-extensions

# 로깅 및 유틸리티
coloredlogs==15.0.1
python-json-logger==2.0.7

# Google Gemini API
google-genai==1.0.0

# 테스트
pytest==7.4.3
httpx==0.25.2 